requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.5
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
setuptools==68.2.2
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
